more work on version extraction
