+++ unlock-js/src/v01/index.js
@@ -1,36 +1,19 @@
-import _getKeyByLockForOwner from './_getKeyByLockForOwner'
-import _getPendingTransaction from './_getPendingTransaction'
-import _getSubmittedTransaction from './_getSubmittedTransaction'
+import { Unlock, PublicLock } from 'unlock-abi-0-1'
 import createLock from './createLock'
-import getKeyByLockForOwner from './getKeyByLockForOwner'
-import getKeysForLockOnPage from './getKeysForLockOnPage'
 import getLock from './getLock'
-import getPastLockCreationsTransactionsForUser from './getPastLockCreationsTransactionsForUser'
-import getPastLockTransactions from './getPastLockTransactions'
-import getTransaction from './getTransaction'
-import getTransactionType from './getTransactionType'
-import parseTransactionFromInput from './parseTransactionFromInput'
 import partialWithdrawFromLock from './partialWithdrawFromLock'
 import purchaseKey from './purchaseKey'
 import updateKeyPrice from './updateKeyPrice'
 import withdrawFromLock from './withdrawFromLock'
 
 export default {
-  _getKeyByLockForOwner,
-  _getPendingTransaction,
-  _getSubmittedTransaction,
   createLock,
-  getKeyByLockForOwner,
-  getKeysForLockOnPage,
   getLock,
-  getPastLockCreationsTransactionsForUser,
-  getPastLockTransactions,
-  getTransaction,
-  getTransactionType,
-  parseTransactionFromInput,
   partialWithdrawFromLock,
   purchaseKey,
   updateKeyPrice,
   withdrawFromLock,
   version: 'v01',
+  Unlock,
+  PublicLock,
 }
